Add a link back to the List view on the project page.
This required a reworking of the `mapDispatchToProps` in App.js. Instead of using Redux's `bindActionCreators`, which I don't really understand, I used a simple function that returns an object that maps the prop we'll use in the Project component, `onSwitchView`, to the function that actually dispatches the action.

See [the section on Container components in the Redux docs](http://redux.js.org/docs/basics/UsageWithReact.html).