Set up UI testing with headless browser.
- Uses [WebdriverIO](http://webdriver.io/), which lets you do Selenium browser automation in NodeJS.
- Run tests with `npm run test`. The Selenium tests run if and when the unit tests pass.
- Currently this just checks to see if the app has a title. Basically an "Are you alive?".
- Uses PhantomJS as the browser.

Requirements:
There needs to be a Selenium server running for this to work. To install a standalone Selenium server:
`curl -O http://selenium-release.storage.googleapis.com/2.53/selenium-server-standalone-2.53.0.jar`

Then, to run it:
`java -jar selenium-server-standalone-2.53.0.jar`

The local server also needs to be running the site at http://localhost:3000/. This is what the tests look at.