Add another way to run an acceptance test.
Run `npm run test-it-the-other-way`. This runs a test using WebdriverIO in "standalone mode". We're just giving this a shot. It may flop. To read up on the difference between the two WebdriverIO modes, [look at this](http://webdriver.io/guide/getstarted/modes.html).

The existing Jest and WebdriverIO tests are still in place and can be run with `npm test`, as before.