:bug: Fix date bug
Addresses MIDASHBRD-137 and MIDASHBRD-139

Normalize the start & end dates when they are received from the server, and before they are placed on the state tree. This gives the date component a date format that it can use, so now the dates show up in the date picker. Also, the time related info (all the stuff after the "T" in the date) is stripped off. We now represent dates on the state tree in YYYY-MM-DD format, and will send them back to the server in this format as well.