Add Status chart
There is a button on the Project page that links to the Status: Demand chart. The chart is using dummy data at this point.

We will need to transform the status data that comes back from the api to a form that the chart can consume. I started building this transform functionality in `helpers/normalizeDemandData`, and started to describe its behavior in a test by the same name. `normalizeDemandData` works for "To Do" so far. We need to make it more general, so that it works for any status ("In Progress", "Done", etc.) See the normalizer in action in `reducers`.

I added a framework for fetching the Status data. The status part of the api is not working for me, so I stubbed in some fake data. (See `actions`.)

To see the stacking functionality of the charts, uncomment the commented lines in `Chart.jsx`.