Use project start & end dates to govern chart range.
One consequence of this is that the app will crash if either of the dates is not defined. Of course, this is bad news. Need to enforce the setting of these dates, somehow...

The solution, for the moment anyway, is to set defaults for project start & end date. This is a somewhat gross, but it keeps the app from crashing in the case that the dates are not set for the project. The date range is intended to span all likely project dates: the default start date is much earlier than any project; the default end date is much later.

In the worst case, the user will look at the Status view, wonder why their charts are all squished up, and then go to the Admin view to set the start and end dates.