Display an S curve on Projection view.
Okay, so it's not exactly an S. Nor is it very curvy. But...

Add Marc's synapse chart and begin adapting:
- Add axes, remove "ticks".
- Redo the point generator function. Pull out into separate file, `makePoints.js`.

Add tests for makePoints.

Add range protections around slider. This keeps the chart from doing weird stuff when it's asked to present impossible/improbable cases. For example, the project is completed within the ramp up period. These cases should be handled within `makePoints`, but for now this makes sure we don't end up with a freaky chart.