Add backlog & dark matter to Projection.
Points maker now incorporates dark matter. (Which means its tests fail. Oh well. We'll have to rewrite them.)

Using global less variables for colors. This allows us to quickly change palettes and it is reflected across the app. Right now, this only affects the projection chart and the color coding in the slider, but can be extended to the Status charts pretty easily, I think.