Modify project chart controls
Make sliders more compact and move the Save button.

The intention is to fit the controls and the chart on the screen without scrolling. So far I've only tested this on my MacBook Pro. We'll eventually need to test for other screen sizes, at some point, but hopefully this provides a good starting point.