:bug: Project completion date displays correctly in Firefox
We're using MomentJS to handle date parsing and formatting. If you don't supply Moment with a date format when parsing a date, it falls back to the native Javascript Date constructor (i.e. `new Date('2016-Apr-08')`). The Date constructor is implemented inconsistently across browsers, so you might find that it works in Chrome and not in Firefox. In fact, this is exactly what we found.

The solution is to tell Moment the format of the date string you want parsed. I did this, and now the date displays correctly in Firefox, Chrome, and Safari.

For reference, see:
https://github.com/moment/moment/issues/1407
http://momentjs.com/guides/#/warnings/js-date/