Filter and normalize the starter projects list
Do some normalization of the api data before sending to components. Specifically, make sure that project ids are represented as strings, not numbers. This type error was causing the app to crash.

I did this normalization following the "selector" pattern favored by Dan Abramov, author of Redux. So all the action happens when we map the state tree to the props, which we then pass to the components. A further step would be to move these selectors/normalizers/filters from mapStateToProps to a more sensible spot.

Also:
- Add a helper to filter by ids.
- Add Mocha/Chai tests for the filterByIds helper.
- Use new filter to filter existing Synapse projects out of the list of starter projects.
- Filter out inactive Harvest projects.
- Give state.project an initial value.