:bug: Fix for incorrect new project initialization
When the user had previously created a new project from an existing Harvest project, and then attempted to create a project "manually" (which means from a blank slate), `blankProject` was overwritten with the data from the previous project.

I'm not sure why this was happening, and it would be good to investigate further. To fix the bug for now, I simply set the starter project to an object literal representing the state of a blank project.