Do not allow "sequence" name values to be blank
Button is disabled until the "Name" field is filled in. Right now, we're just checking for the presence of any value. We could easily make the validation more sophisticated, such as checking for three or more characters, or whatever logic we like.

I switched to handling the state of this miniature form using the component's state instead of ref. To do so, I switched to using the class syntax instead of the pure function deal.