Add demand status data validity checker
This checks the status data that comes from the api. The data must meet certain criteria before we pass it to the chart to be rendered.

For all status data, for all three types (demand, defect, and effort):
- Check validity of keys
- Check validity of dates

If the data is invalid show the user a message explaining what happened.

Test coverage for several cases is included.