SYNAPSE-5 #comment Status chart styling
The effort chart should be a line chart, and the other two should be stacked area charts. Right now we are using one function to generate all three: "renderStackedAreaChart". This commit addresses this issue... sort of. It updates the way the effort chart is styled so that we give the effort chart the appearance of a line chart.

This "works", but structurally the effort chart is still a stacked area chart. We really should use a separate function that generates a proper line chart.

I also styled the legend to match the previous style: all caps, etc.