Reproduce Eolas's test data script on the client side
We have a script in Eolas that generates test data. This commit reproduces the creation of that test data, but on the client side. See the README for more on usage instructions.

To make this actually work, we need to redo the various AJAX calls so that they coerce the incoming data to JSON. See the "fetchProjects" action for a working example.

Rather than redo all these actions now, I decided to wait until the actions refactor is complete. We can fold this in at that time.