Add function for getting the value of demand data at any given date
The projection alarm relies on this function. It should have been included in a previous commit, but got left out somehow. Probably my mistake.

Note that this function assumes that the projection line is a series of four lines that can be described by simple linear functions. If we change the shape of the projection line, this function will need to be rewritten.