:bug: All status data is now visible
For some projects (e.g. Test Project 1) the status data was truncated at the right end of the chart. Status data, you are truncated no more.

This commit fixes the immediate issue, but at some point we'll need to review the usage of all the values in renderStatusChart's "config" file. As is, there are a few things that don't seem to be working quite as you'd expect.

I also fixed a few other little visual crufts: the scrubber line was too long, the projection "alarm" interfered with the scrubber line.