Swap precommit packages
This new package is the one Shahzain recommends. Hoping this will fix the issue where an additional "precommit" key is automatically added to package.json.

Otherwise, it should work almost exactly as before. Whenever you make a commit, it runs the linter, tests, etc. The only difference is this: where the previous precommit only fired on the master branch, this one fires on any commit. If you want to skip the hook (if, for instance, you're hacking away on a local branch and want to make WIP commits) you can use the `--no-verify` flag.