SYNAPSE-80 #comment Predict project completion date
After some experimentation, I decided to go with a second order polynomial regression to calculate the forecasted end date. The status data is the input, and the output is a quadratic forumula like this:

y = ax^2 + bx + c

From here, we can solve for y = 0, which gives the forecasted end date of the project. We can also plot this curve, although I'm not so sure it makes sense to do so on top of a stacked area chart. The second order regression seems to fit a typical Backlog curve. We can certainly revisit this if we encounter some crazy backlog data.

This commit shows the forecasted end date on the Status page. It also draws the regression curve on the Demand cumulative flow diagram (again, this may be more confusing than clarifying).

It's worth noting that these regression calculations are independent of the chart itself. This means we could do the calculation anywhere in the app. It wouldn't be too much of a stretch to display the forecasted end date of each project in the home screen.