SYNAPSE-80 #comment Add toggle for showing regression curve
We don't want to show the regression curve to the user. It's weird and confusing. But it is helpful to visualize the math while coding for various scenarios. To facilitate this there's now a toggle for showing/hiding the regression curve. At the top of renderStatusChart/index set `showRegressionCurve` to true and the curve will appear overlaid on the demand Status chart.