SYNAPSE-80 #comment Display forecasted completion date and linear regression lines for Done and Backlog.
The linear regression is calculated using data only over the Done date range. The idea here is that any activity before "Done" is discovery and should not be considered when calculating the trends.
The completion date is assumed to be the date where Backlog equals 0.
If the completion date cannot be calculated (e.g. Backlog trends upward infinitely) hide the regression lines and display an appropriate message to the user.