SYNAPSE-93 #comment :bug: Fix Projection view bug
The app was crashing because projection was undefined on the state tree in some cases. I added a conditional statement to check for the presence of the projection. If it doesn't exist, then it uses default values for the projection.

This solution got the Projection view up and running again, but it feels like we need to look more deeply at the issue. This validation/default value stuff should be happening further up the data flow. I'm not sure why it's not.