Add Login component
The auth itself if faked out for the moment. Can log in with Paul as name. Blocks any other name. Doesn't care about password.

On the Project List view, hide "New" button when the user is logged out.
On the Project View page, hide the buttons when the user is logged out.
Modify acceptance tests so they are not dependent on these buttons.