Add "Fixer Uppers", which validate and normalize data from APIs
There are two layers to the fixerUpper: validation and normalization. The idea is that the validation errors drive the normalizer. For example, say the API serves data with a date formatted "MM/DD/YYYY" and our schema says it should be formatted "MM-DD-YYYY". The validator will issue the complaint, and the normalizer will use this complaint as the basis for doing the cleanup. Right now the only place where this is really happening is with the projection; in most cases we're just logging the validation errors.

The structure is there for running all API responses through this "fixerUpper" layer. Currently, only the project data is running through this process. Follow the pattern of the project fixerUpper when adding additional fixerUppers.

I also made these changes:
- Do the fixing up of project data (which includes normalization) in the reducers instead of the components
- Added a project schema, which is used by the project validator.