SYNAPSE-89 #comment Refresh no longer logs out user
If the user has already logged in, they get logged in automatically, even if they've shut down the browser in the meantime. The app is doing a simple check for the presence of a "user" in local storage. So this is really not secure! But I think it gets us to a "good enough" sort of auth for now.