SYNAPSE-91 #comment Add missing defect fields
These fields were already there. I just made them visible on the Project View component.

I also fixed a typo in one of the keys. This was causing the property to be saved under the wrong key.

The story in JIRA says the fields should be called initialStatus and resolvedStatus. I left them with their existing names, not wanting t break the current API. But, of course, these key names could be easily changed.